fix: IBD build fixes + script/block/checkqueue optimizations
- block: BIP143 refs borrow fix — use local refs Vec in bip143 block, drop refs_buf (E0502)
- script: BATCH_PUT_SIG_CACHE_BY_SHARD .with() closure; SendSyncUnsafeCell for EcdsaSoAStorage; stack borrow fixes (pass stack not &mut stack)
- checkqueue: WORKER_PV_BUF .with() for thread_local; TX_TO_INDICES grouping; batch_results per batch
- ScriptCheckQueue, script_profile, block production path: refs reuse, script_pubkey_vec, prevout_values
- IBD optimizations (4b–4d, 5a, 6a–6d): du lock split, feeder bytes, mi_collect adaptive, etc.

Made-with: Cursor

+++ benches/bllvm_optimizations.rs
@@ -55,7 +55,7 @@ fn create_test_transactions(count: usize) -> Vec<Transaction> {
     (0..count)
         .map(|i| {
             let mut tx = create_test_transaction();
-            tx.inputs[0].prevout.index = i as u64;
+            tx.inputs[0].prevout.index = i as u32;
             tx
         })
         .collect()
